Separate space charge kick elements into two separate files
