+++ cheetah/accelerator/__init__.py
@@ -14,8 +14,8 @@ from .screen import Screen  # noqa: F401
 from .segment import Segment  # noqa: F401
 from .sextupole import Sextupole  # noqa: F401
 from .solenoid import Solenoid  # noqa: F401
-from .space_charge_kick import SpaceChargeKick2D  # noqa: F401
-from .space_charge_kick import SpaceChargeKick3D  # noqa: F401
+from .space_charge_kick_2d import SpaceChargeKick2D  # noqa: F401
+from .space_charge_kick_3d import SpaceChargeKick3D  # noqa: F401
 from .transverse_deflecting_cavity import TransverseDeflectingCavity  # noqa: F401
 from .undulator import Undulator  # noqa: F401
 from .vertical_corrector import VerticalCorrector  # noqa: F401
